+++ packages/render-composer/src/Layers.ts
@@ -1,0 +1,4 @@
+export const Layers = {
+  Default: 0,
+  TransparentFX: 1
+}
