+++ packages/render-composer/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./Canvas"
 export * from "./CopyPass"
-export * from "./DefaultRenderPipeline"
 export * from "./DepthCopyPass"
 export * from "./EffectComposer"
 export * from "./EffectPass"
@@ -10,3 +9,4 @@ export * from "./RenderPipeline"
 export * from "./SelectiveBloomEffect"
 export * from "./SMAAEffect"
 export * from "./VignetteEffect"
+export * from "./Layers"
